Fix some implicit-widening-of-multiplication-result warnings See https://clang.llvm.org/extra/clang-tidy/checks/bugprone/implicit-widening-of-multiplication-result.html
I applied the following rules of thumb:
- use point_cloud.resize(width, height) instead of point_cloud.resize(width*height)
- cast factors to `std::size_t` if multiplication result would be cast to `std::size_t` later anyway (e.g. used for vector.resize, vector.reserve, vector[])
- change some variables to type `std::size_t` if they otherwise had to be cast repeatedly

We actually had some reports in the past where people were working with large point types (at least 32 bytes per point) and a large number of points, so that (bytes per point)*(height)*(width) would overflow if done in 32 bit types (e.g. when accessing the binary data in PCLPointCloud2). Fixing the implicit-widening-of-multiplication-result warnings would prevent such kinds of problems.

+++ common/include/pcl/range_image/impl/range_image.hpp
@@ -194,7 +194,7 @@ RangeImage::createFromPointCloudWithKnownSize (const PointCloudType& point_cloud
   image_offset_y_ = (std::max) (0, center_pixel_y-pixel_radius_y);
 
   points.clear ();
-  points.resize (width*height, unobserved_point);
+  points.resize (static_cast<std::size_t>(width)*static_cast<std::size_t>(height), unobserved_point);
   
   int top=height, right=-1, bottom=-1, left=width;
   doZBuffer (point_cloud, noise_level, min_range, top, right, bottom, left);
